Phase 4: extract i18n, split chat UI, unify email, add E2E
- Move translations to locales/en.ts and locales/fr.ts (~180 line context)
- Split chat-interface into components/chat/* and use-message-search hook
- Add lib/email barrel (campaigns, transactional, client)
- Remove orphan sign-in-new, sign-up-new, and OpenClaw UI paths
- Add Playwright smoke tests (homepage, auth, chat, health, 404 checks)
- Configure vitest to exclude e2e directory
