+++ .gitignore
@@ -10,6 +10,10 @@
 # production
 /build
 
+# test artifacts
+/playwright-report/
+/test-results/
+
 # debug
 npm-debug.log*
 yarn-debug.log*
