+++ locales/en.ts
@@ -1,0 +1,499 @@
+import type { TranslationStrings } from "./types"
+
+export const en: TranslationStrings = {
+    // Common
+    "app.name": "PocketLawyer",
+    "app.tagline": "Your Personal Legal Assistant for Cameroonian Law",
+    "app.copyright": "© 2025 PocketLawyer. All rights reserved.",
+
+    // Auth
+    "auth.signin": "Log in",
+    "auth.signup": "Create Account",
+    "auth.signin.title": "Welcome back",
+    "auth.signup.title": "Join PocketLawyer",
+    "auth.signin.description": "Log in to your account",
+    "auth.signup.description": "Create your free account",
+    "auth.signin.google": "Continue with Google",
+    "auth.signup.google": "Continue with Google",
+    "auth.signin.email": "Continue with Email",
+    "auth.signup.email": "Continue with Email",
+    "auth.or": "or",
+    "auth.email": "Your email",
+    "auth.password": "Your password",
+    "auth.confirm.password": "Confirm password",
+    "auth.name": "Your name",
+    "auth.forgot.password": "Forgot your password?",
+    "auth.no.account": "Don't have an account?",
+    "auth.has.account": "Already have an account?",
+    "auth.creating": "Creating your account...",
+    "auth.signing.in": "Logging you in...",
+    "auth.signin.button": "Log in",
+    "auth.orContinueWith": "or",
+    "auth.continueWithGoogle": "Continue with Google",
+    "auth.noAccount": "New to PocketLawyer?",
+    "auth.createAccount": "Create an account",
+
+    // Welcome page
+    "welcome.hero.title": "Your AI Legal Assistant for Cameroonian Law",
+    "welcome.hero.subtitle":
+      "Get instant answers to your legal questions, browse relevant resources, and access personalized legal guidance powered by AI.",
+    "welcome.get.started": "Get Started",
+    "welcome.features.title": "Key Features",
+    "welcome.feature.chat.title": "AI Chat Assistant",
+    "welcome.feature.chat.description":
+      "Get instant answers to your legal questions with our specialized AI trained on Cameroonian law.",
+    "welcome.feature.search.title": "Web Search Integration",
+    "welcome.feature.search.description":
+      "Browse relevant legal resources and find specific information about Cameroonian law.",
+    "welcome.feature.personalized.title": "Personalized Experience",
+    "welcome.feature.personalized.description":
+      "Create an account to save your chat history and get more personalized legal guidance.",
+    "welcome.feature.document.title": "Document Analysis",
+    "welcome.feature.document.description":
+      "Upload legal documents for AI-powered analysis and get insights specific to your documents.",
+    "welcome.understand.title": "Understand Cameroonian Law",
+    "welcome.understand.description":
+      "Cameroon's legal system is unique, combining elements of English Common Law, French Civil Law, and customary law. Our AI assistant helps you navigate this complex legal landscape.",
+    "welcome.understand.point1": "Access information on civil, criminal, and commercial law",
+    "welcome.understand.point2": "Learn about your rights and legal procedures",
+    "welcome.understand.point3": "Get guidance on legal documents and requirements",
+    "welcome.justice.title": "Justice Accessible to All",
+    "welcome.justice.description": "Our mission is to make legal information accessible to everyone in Cameroon",
+    "welcome.ready.title": "Ready to get started?",
+    "welcome.ready.description": "Sign up now to access all features and get personalized legal assistance.",
+    "welcome.create.account": "Create Your Free Account",
+    "welcome.footer.terms": "Terms",
+    "welcome.footer.privacy": "Privacy",
+    "welcome.footer.contact": "Contact",
+
+    // Chat interface
+    "chat.tab.chat": "Chat",
+    "chat.tab.web": "Web Search",
+    "chat.tab.document": "Document",
+    "chat.welcome.title": "Welcome to PocketLawyer",
+    "chat.welcome.subtitle": "Ask any legal question about Cameroonian law and get helpful answers.",
+    "chat.try.asking": "Try asking:",
+    "chat.example1": "What are the requirements for starting a business in Cameroon?",
+    "chat.example2": "Explain the divorce process in Cameroon",
+    "chat.example3": "What are my rights if I'm arrested in Cameroon?",
+    "chat.signin.prompt": "Sign in for unlimited messages and personalized history",
+    "chat.message.limit": "Message limit reached. Sign in to continue chatting.",
+    "chat.messages.remaining": "{count} messages remaining. Sign in for unlimited access.",
+    "chat.input.placeholder": "Ask a legal question...",
+    "chat.send": "Send",
+    "chat.search.placeholder": "search", // Updated placeholder
+    "chat.search": "Search",
+
+    // Chat errors
+    "chat.not.found": "Chat not found. Redirecting to new chat...",
+    "chat.unauthorized": "You don't have access to this chat. Redirecting to new chat...",
+    "chat.error.loading": "Error loading chat. Please try again.",
+
+    // Document analysis
+    "document.select": "Select Document",
+    "document.signin.prompt": "Sign in for enhanced document analysis and to save your results",
+    "document.preview": "Document Preview",
+    "document.question.prompt": "Ask a question about this document",
+    "document.question.placeholder":
+      "E.g., What are the key legal points in this document? Or, Can you summarize this document?",
+    "document.analyze": "Analyze Document",
+    "document.analyzing": "Analyzing...",
+
+    // Quick Actions
+    "Quick Actions": "Quick Actions",
+    "Analyze Document": "Analyze Document",
+    "Draft Contract": "Draft Contract",
+    "Legal Research": "Legal Research",
+    "Case Review": "Case Review",
+    "Find a Lawyer": "Find a Lawyer",
+    "Book Consultation": "Book Consultation",
+    "Contract drafting feature coming soon": "Contract drafting feature coming soon",
+    "Legal research feature coming soon": "Legal research feature coming soon",
+    "Case review feature coming soon": "Case review feature coming soon",
+
+    // Sidebar
+    "sidebar.new.chat": "New Chat",
+    "sidebar.no.history": "No chat history yet",
+    "sidebar.start.conversation": "Start a new conversation to see it here",
+    "sidebar.delete.chat": "Delete Chat",
+    "sidebar.delete.confirm": "Are you sure you want to delete this chat? This action cannot be undone.",
+    "sidebar.cancel": "Cancel",
+    "sidebar.delete": "Delete",
+    "sidebar.signed.with.google": "Signed in with Google",
+    "sidebar.signout": "Sign out",
+
+    // Language
+    "language.en": "English",
+    "language.fr": "French",
+    "language.switch": "Switch to {language}",
+
+    // Profile
+    "profile.title": "Profile Settings",
+    "profile.description": "Update your profile information and password",
+    "profile.tab": "Profile",
+    "profile.password.tab": "Password",
+    "profile.name": "Name",
+    "profile.email": "Email",
+    "profile.email.google": "Email cannot be changed for Google accounts",
+    "profile.update": "Update Profile",
+    "profile.updated": "Profile updated successfully",
+    "profile.current.password": "Current Password",
+    "profile.new.password": "New Password",
+    "profile.confirm.password": "Confirm New Password",
+    "profile.password.mismatch": "New passwords don't match",
+    "profile.password.updated": "Password updated successfully",
+    "profile.password.google": "Password management is handled by Google for Google accounts",
+    "profile.update.password": "Update Password",
+
+    // Password Reset
+    "auth.reset.password.title": "Reset Password",
+    "auth.reset.password.description": "Enter your email address and we'll send you a password reset link",
+    "auth.reset.password.button": "Send Reset Link",
+    "auth.reset.password.success": "Password reset link sent! Check your email",
+
+    // Feedback
+    "feedback.button": "Give Feedback",
+    "feedback.title": "Help us improve",
+    "feedback.rating.required": "Please select a rating",
+    "feedback.comment.placeholder": "Share your experience (optional)",
+    "feedback.submit": "Submit Feedback",
+    "feedback.submitting": "Submitting...",
+    "feedback.success": "Thank you for your feedback!",
+    "feedback.error": "Failed to submit feedback. Please try again.",
+
+    // Contact page
+    "contact.title": "Contact Us",
+    "contact.description": "Have questions or feedback? Send us a message below.",
+    "contact.name": "Name",
+    "contact.email": "Email",
+    "contact.message": "Message",
+    "contact.send": "Send Message",
+    "contact.sending": "Sending...",
+    "contact.back": "Back to Home",
+    "contact.success": "Your message has been sent successfully!",
+    "contact.error": "Failed to send message. Please try again later.",
+
+    // Not found page
+    "notfound.title": "Page Not Found",
+    "notfound.description": "The page you're looking for doesn't exist or has been moved.",
+    "notfound.go.home": "Go Home",
+    "notfound.find.lawyers": "Find Lawyers",
+
+    // Onboarding loading
+    "onboarding.loading": "Loading onboarding...",
+
+    // Profile page (extended)
+    "profile.settings.title": "Profile Settings",
+    "profile.settings.description": "Update your profile information and password",
+    "profile.picture": "Profile Picture",
+    "profile.picture.updated": "Profile picture updated successfully!",
+    "profile.picture.failed": "Failed to save profile picture",
+    "profile.update.button": "Update Profile",
+    "profile.update.password.button": "Update Password",
+    "profile.go.to.admin": "Go To Admin",
+
+    // Lawyer profile edit page
+    "lawyer.edit.back": "Back to Dashboard",
+    "lawyer.edit.title": "Edit Profile",
+    "lawyer.edit.subtitle": "Update your professional information",
+    "lawyer.edit.picture": "Profile Picture",
+    "lawyer.edit.picture.desc": "Upload a professional photo (max 2MB)",
+    "lawyer.edit.basic": "Basic Information",
+    "lawyer.edit.fullname": "Full Name *",
+    "lawyer.edit.email": "Email *",
+    "lawyer.edit.phone": "Phone Number *",
+    "lawyer.edit.bar": "Bar Number",
+    "lawyer.edit.bio": "Professional Bio *",
+    "lawyer.edit.bio.placeholder": "Tell potential clients about yourself...",
+    "lawyer.edit.professional": "Professional Details",
+    "lawyer.edit.specialties": "Specialties * (Select all that apply)",
+    "lawyer.edit.experience": "Years of Experience *",
+    "lawyer.edit.rate": "Hourly Rate (XAF) *",
+    "lawyer.edit.education": "Education",
+    "lawyer.edit.languages": "Languages",
+    "lawyer.edit.add.language": "Add language",
+    "lawyer.edit.add": "Add",
+    "lawyer.edit.save": "Save Changes",
+    "lawyer.edit.saving": "Saving...",
+    "lawyer.edit.cancel": "Cancel",
+    "lawyer.edit.error.name": "Name is required",
+    "lawyer.edit.error.phone": "Phone is required",
+    "lawyer.edit.error.bio": "Bio is required",
+    "lawyer.edit.error.specialty": "Select at least one specialty",
+    "lawyer.edit.error.experience": "Valid experience is required",
+    "lawyer.edit.error.rate": "Valid hourly rate is required",
+    "lawyer.edit.success": "Profile updated successfully!",
+    "lawyer.edit.load.error": "Failed to load profile",
+    "lawyer.edit.save.error": "Failed to update profile",
+    "lawyer.edit.not.found": "Lawyer profile not found",
+
+    // Lawyer dashboard
+    "lawyer.dashboard.pending": "PENDING",
+    "lawyer.dashboard.upcoming": "UPCOMING",
+    "lawyer.dashboard.history": "HISTORY",
+    "lawyer.dashboard.config": "CONFIG",
+    "lawyer.dashboard.no.pending": "No pending bookings",
+    "lawyer.dashboard.no.upcoming": "No upcoming bookings",
+    "lawyer.dashboard.no.completed": "No completed bookings yet",
+    "lawyer.dashboard.availability": "AVAILABILITY CONFIG",
+    "lawyer.dashboard.availability.desc": "Configure your weekly schedule",
+    "lawyer.dashboard.profile.settings": "PROFILE SETTINGS",
+    "lawyer.dashboard.profile.desc": "Manage your professional profile",
+    "lawyer.dashboard.account": "ACCOUNT",
+    "lawyer.dashboard.edit.profile": "EDIT PROFILE",
+    "lawyer.dashboard.client.view": "CLIENT VIEW",
+    "lawyer.dashboard.sign.out": "SIGN OUT",
+    "lawyer.dashboard.edit": "EDIT",
+    "lawyer.dashboard.offline": "OFFLINE",
+    "lawyer.dashboard.slots": "{count} slots",
+    "lawyer.dashboard.confirm.title": "CONFIRM BOOKING",
+    "lawyer.dashboard.confirm.desc": "Approve this consultation request",
+    "lawyer.dashboard.confirm.meeting": "Meeting Link",
+    "lawyer.dashboard.confirm.btn": "CONFIRM",
+    "lawyer.dashboard.cancel.title": "CANCEL BOOKING",
+    "lawyer.dashboard.cancel.desc": "This action cannot be undone",
+    "lawyer.dashboard.abort": "ABORT",
+    "lawyer.dashboard.cancel.btn": "CANCEL BOOKING",
+    "lawyer.dashboard.avail.title": "EDIT AVAILABILITY",
+    "lawyer.dashboard.avail.desc": "Configure your weekly schedule",
+    "lawyer.dashboard.save": "SAVE",
+    "lawyer.dashboard.cancel": "CANCEL",
+    "lawyer.dashboard.confirm.booking": "CONFIRM",
+    "lawyer.dashboard.decline": "DECLINE",
+    "lawyer.dashboard.complete": "COMPLETE",
+    "lawyer.dashboard.initializing": "Initializing dashboard...",
+    "lawyer.dashboard.label.pending": "PENDING",
+    "lawyer.dashboard.label.upcoming": "UPCOMING",
+    "lawyer.dashboard.label.completed": "COMPLETED",
+    "lawyer.dashboard.label.earnings": "EARNINGS",
+    "lawyer.dashboard.label.rating": "RATING",
+    "lawyer.dashboard.label.reviews": "REVIEWS",
+    "lawyer.dashboard.label.rate": "HOURLY RATE",
+    "lawyer.dashboard.label.status": "STATUS",
+    "lawyer.dashboard.label.specialties": "SPECIALTIES",
+    "lawyer.dashboard.label.experience": "EXPERIENCE",
+    "lawyer.dashboard.years": "{n} years",
+    "lawyer.dashboard.exit": "EXIT",
+
+    // Footer disclaimer
+    "footer.disclaimer": "PocketLawyer provides general legal information based on Cameroonian law. It does not offer personalized legal advice, cannot represent you in court, and should not replace consultation with a qualified attorney.",
+    "footer.capabilities": "What PocketLawyer can and cannot do",
+    "footer.can": "PocketLawyer CAN:",
+    "footer.cannot": "PocketLawyer CANNOT:",
+    "footer.can.1": "Provide general information about Cameroonian law",
+    "footer.can.2": "Explain legal concepts and terminology",
+    "footer.can.3": "Point to relevant statutes and legal resources",
+    "footer.can.4": "Help understand legal procedures and requirements",
+    "footer.can.5": "Generate basic document templates",
+    "footer.cannot.1": "Provide personalized legal advice for specific situations",
+    "footer.cannot.2": "Represent users in court or legal proceedings",
+    "footer.cannot.3": "File documents on behalf of users",
+    "footer.cannot.4": "Guarantee outcomes in legal matters",
+    "footer.cannot.5": "Replace professional legal counsel",
+    "footer.terms": "Terms of Service",
+    "footer.privacy": "Privacy Policy",
+
+    // Common shared strings
+    "Home": "Home",
+    "Back": "Back",
+    "Cancel": "Cancel",
+    "Copy": "Copy",
+    "Copied": "Copied",
+    "Download": "Download",
+    "Save": "Save",
+    "optional": "optional",
+    "Blog": "Blog",
+    "Examples": "Examples",
+    "Return to Chat": "Return to Chat",
+    "Copied to clipboard": "Copied to clipboard",
+    "Documents": "Documents",
+    "Chat": "Chat",
+
+    // Blog page
+    "Legal Resources & Guides": "Legal Resources & Guides",
+    "Articles, guides and resources on Cameroonian law to help you understand your rights and obligations.": "Articles, guides and resources on Cameroonian law to help you understand your rights and obligations.",
+    "Search articles...": "Search articles...",
+    "View Examples": "View Examples",
+    "Loading articles...": "Loading articles...",
+    "Failed to load blog posts. Please try again later.": "Failed to load blog posts. Please try again later.",
+    "Try Again": "Try Again",
+    "Categories": "Categories",
+    "All Categories": "All Categories",
+    "Popular Tags": "Popular Tags",
+    "Have a Legal Question?": "Have a Legal Question?",
+    "Use our AI assistant to get instant answers to your Cameroonian legal questions.": "Use our AI assistant to get instant answers to your Cameroonian legal questions.",
+    "Chat with PocketLawyer": "Chat with PocketLawyer",
+    "No articles found": "No articles found",
+    "Try adjusting your search or category filters": "Try adjusting your search or category filters",
+    "Reset Filters": "Reset Filters",
+    "Subscribe to Our Legal Updates": "Subscribe to Our Legal Updates",
+    "Get the latest articles and resources on Cameroonian law delivered to your inbox": "Get the latest articles and resources on Cameroonian law delivered to your inbox",
+    "Enter your email": "Enter your email",
+    "Subscribing...": "Subscribing...",
+    "Subscribe": "Subscribe",
+    "Please enter your email address": "Please enter your email address",
+    "Failed to process subscription. Please try again.": "Failed to process subscription. Please try again.",
+
+    // Draft Contract page
+    "Generate professional legal contracts compliant with Cameroonian law and OHADA": "Generate professional legal contracts compliant with Cameroonian law and OHADA",
+    "AI-powered contract generation for Cameroon": "AI-powered contract generation for Cameroon",
+    "Select Contract Type": "Select Contract Type",
+    "Employment Contract": "Employment Contract",
+    "CDI/CDD under Cameroon Labour Code": "CDI/CDD under Cameroon Labour Code",
+    "Lease Agreement": "Lease Agreement",
+    "Residential or commercial rental": "Residential or commercial rental",
+    "Sale of Goods": "Sale of Goods",
+    "Commercial sale under OHADA": "Commercial sale under OHADA",
+    "Service Agreement": "Service Agreement",
+    "Consulting or professional services": "Consulting or professional services",
+    "Non-Disclosure Agreement": "Non-Disclosure Agreement",
+    "Confidentiality agreement": "Confidentiality agreement",
+    "Partnership Agreement": "Partnership Agreement",
+    "Business partnership terms": "Business partnership terms",
+    "Loan Agreement": "Loan Agreement",
+    "Personal or business loan": "Personal or business loan",
+    "Power of Attorney": "Power of Attorney",
+    "Procuration / legal representation": "Procuration / legal representation",
+    "Custom Contract": "Custom Contract",
+    "Describe your specific needs": "Describe your specific needs",
+    "Additional Requirements": "Additional Requirements",
+    "Describe any specific requirements: parties involved, key terms, special clauses, language preference (English/French)...": "Describe any specific requirements: parties involved, key terms, special clauses, language preference (English/French)...",
+    "Generate Contract": "Generate Contract",
+    "New Contract": "New Contract",
+    "Drafting contract...": "Drafting contract...",
+    "Request modifications to the contract...": "Request modifications to the contract...",
+    "Contracts generated using Cameroonian law templates and OHADA provisions": "Contracts generated using Cameroonian law templates and OHADA provisions",
+    "Contract downloaded": "Contract downloaded",
+    "Please select a contract type": "Please select a contract type",
+
+    // Bookings page
+    "My Bookings": "My Bookings",
+    "Manage your legal consultations": "Manage your legal consultations",
+    "Upcoming Consultations": "Upcoming Consultations",
+    "Past Consultations": "Past Consultations",
+    "Total Bookings": "Total Bookings",
+    "Upcoming": "Upcoming",
+    "Past": "Past",
+    "No upcoming consultations": "No upcoming consultations",
+    "Browse Lawyers": "Browse Lawyers",
+    "No past consultations": "No past consultations",
+    "Cancel Booking": "Cancel Booking",
+    "Please provide a reason for cancelling this consultation": "Please provide a reason for cancelling this consultation",
+    "Cancellation Reason": "Cancellation Reason",
+    "e.g., Schedule conflict, found another lawyer...": "e.g., Schedule conflict, found another lawyer...",
+    "Keep Booking": "Keep Booking",
+    "Cancelling...": "Cancelling...",
+    "Leave a Review": "Leave a Review",
+    "Share your experience with": "Share your experience with",
+    "Rating": "Rating",
+    "Your Review": "Your Review",
+    "Share your thoughts about the consultation...": "Share your thoughts about the consultation...",
+    "Submitting...": "Submitting...",
+    "Submit Review": "Submit Review",
+    "minutes": "minutes",
+    "Notes": "Notes",
+    "Join Video Call": "Join Video Call",
+    "Cancelled by": "Cancelled by",
+    "Total": "Total",
+    "Payment": "Payment",
+    "Leave Review": "Leave Review",
+    "Review submitted": "Review submitted",
+    "Failed to load bookings": "Failed to load bookings",
+    "Please sign in to cancel bookings": "Please sign in to cancel bookings",
+    "Please provide a cancellation reason": "Please provide a cancellation reason",
+    "Booking cancelled successfully": "Booking cancelled successfully",
+    "Failed to cancel booking": "Failed to cancel booking",
+    "Please sign in to submit reviews": "Please sign in to submit reviews",
+    "Please write a review comment": "Please write a review comment",
+    "Review submitted successfully!": "Review submitted successfully!",
+    "Failed to submit review": "Failed to submit review",
+    "Pending": "Pending",
+    "Confirmed": "Confirmed",
+    "Completed": "Completed",
+    "Cancelled": "Cancelled",
+
+    // Case Review page
+    "Get an AI-powered analysis of your legal case under Cameroonian law": "Get an AI-powered analysis of your legal case under Cameroonian law",
+    "AI analysis of legal cases under Cameroonian law": "AI analysis of legal cases under Cameroonian law",
+    "Disclaimer": "Disclaimer",
+    "This tool provides an AI-generated analysis for informational purposes only. It is not legal advice. Please consult a qualified Cameroonian lawyer for professional legal counsel.": "This tool provides an AI-generated analysis for informational purposes only. It is not legal advice. Please consult a qualified Cameroonian lawyer for professional legal counsel.",
+    "Case Type": "Case Type",
+    "Select case type...": "Select case type...",
+    "Civil Dispute": "Civil Dispute",
+    "Criminal Matter": "Criminal Matter",
+    "Commercial / Business": "Commercial / Business",
+    "Labour / Employment": "Labour / Employment",
+    "Land / Property": "Land / Property",
+    "Family Law": "Family Law",
+    "Administrative": "Administrative",
+    "Other": "Other",
+    "Describe Your Case": "Describe Your Case",
+    "Describe the facts of your case in detail. Include: what happened, when, who was involved, any documents or agreements, and what outcome you are seeking...": "Describe the facts of your case in detail. Include: what happened, when, who was involved, any documents or agreements, and what outcome you are seeking...",
+    "Example Cases": "Example Cases",
+    "Wrongful Dismissal": "Wrongful Dismissal",
+    "Employee fired without notice after 5 years of service": "Employee fired without notice after 5 years of service",
+    "Land Dispute": "Land Dispute",
+    "Neighbour building on my registered land": "Neighbour building on my registered land",
+    "Unpaid Invoice": "Unpaid Invoice",
+    "Client refuses to pay for delivered goods": "Client refuses to pay for delivered goods",
+    "Use this example": "Use this example",
+    "Start Case Review": "Start Case Review",
+    "New Case Review": "New Case Review",
+    "Analysing case...": "Analysing case...",
+    "Ask a follow-up question about your case...": "Ask a follow-up question about your case...",
+    "Analysis enhanced with built-in Cameroonian law knowledge base": "Analysis enhanced with built-in Cameroonian law knowledge base",
+    "Please describe your case details": "Please describe your case details",
+
+    // Legal Research page
+    "Research Cameroonian law with AI-powered assistance and built-in legal knowledge base": "Research Cameroonian law with AI-powered assistance and built-in legal knowledge base",
+    "Powered by Cameroonian law knowledge base + AI": "Powered by Cameroonian law knowledge base + AI",
+    "Popular Research Topics": "Popular Research Topics",
+    "Land & Property Law": "Land & Property Law",
+    "Business Formation": "Business Formation",
+    "Labour & Employment": "Labour & Employment",
+    "Criminal Law": "Criminal Law",
+    "Tax Law": "Tax Law",
+    "Researching...": "Researching...",
+    "Describe the legal topic you want to research...": "Describe the legal topic you want to research...",
+    "Responses enhanced with built-in Cameroonian law knowledge base (available offline)": "Responses enhanced with built-in Cameroonian law knowledge base (available offline)",
+
+    // Documents page
+    "Document Analysis Center": "Document Analysis Center",
+    "Upload, analyze, and extract insights from your legal documents": "Upload, analyze, and extract insights from your legal documents",
+    "Processed Documents": "Processed Documents",
+    "Click on a document to analyze it": "Click on a document to analyze it",
+    "Select a Document": "Select a Document",
+    "characters": "characters",
+    "Document Preview": "Document Preview",
+    "Quick Analysis": "Quick Analysis",
+    "Summarize the key points of this document": "Summarize the key points of this document",
+    "Who are the parties involved and what are their obligations?": "Who are the parties involved and what are their obligations?",
+    "What are the important dates and deadlines?": "What are the important dates and deadlines?",
+    "Identify any potential legal risks or problematic clauses": "Identify any potential legal risks or problematic clauses",
+    "What legal rights are established in this document?": "What legal rights are established in this document?",
+    "Is this document enforceable under Cameroonian law?": "Is this document enforceable under Cameroonian law?",
+    "Custom Question": "Custom Question",
+    "Ask a specific question about this document...": "Ask a specific question about this document...",
+    "Tip": "Tip",
+    "to submit": "to submit",
+    "Analyzing…": "Analyzing…",
+    "Analyze": "Analyze",
+    "Analyzing your document with AI…": "Analyzing your document with AI…",
+    "Analysis History": "Analysis History",
+    "Document was truncated. Analysis based on first portion only.": "Document was truncated. Analysis based on first portion only.",
+    "Question": "Question",
+    "Answer": "Answer",
+    "No document selected": "No document selected",
+    "Upload a document above or select one from the list to start analyzing": "Upload a document above or select one from the list to start analyzing",
+    "Quick Tips": "Quick Tips",
+    "Supported Formats": "Supported Formats",
+    "PDF, DOCX, and TXT files up to 10MB each": "PDF, DOCX, and TXT files up to 10MB each",
+    "OCR Support": "OCR Support",
+    "Scanned PDFs automatically fall back to Tesseract OCR": "Scanned PDFs automatically fall back to Tesseract OCR",
+    "AI Analysis": "AI Analysis",
+    "Ask specific questions about your documents for detailed insights": "Ask specific questions about your documents for detailed insights",
+    "Secure Processing": "Secure Processing",
+    "All documents are processed securely and confidentially": "All documents are processed securely and confidentially",
+    "Please select a document and enter a question": "Please select a document and enter a question",
+    "Analysis failed": "Analysis failed",
+}
