+++ locales/index.ts
@@ -1,0 +1,7 @@
+import { en } from "./en";
+import { fr } from "./fr";
+import type { Translations } from "./types";
+
+export const translations: Translations = { en, fr };
+
+export type { TranslationStrings, Translations } from "./types";
