+++ locales/types.ts
@@ -1,0 +1,6 @@
+export type TranslationStrings = Record<string, string>;
+
+export type Translations = {
+  en: TranslationStrings;
+  fr: TranslationStrings;
+};
