+++ package.json
@@ -8,7 +8,9 @@
     "build:next": "next build",
     "start": "next start",
     "lint": "next lint",
-    "test": "vitest run"
+    "test": "vitest run",
+    "test:e2e": "playwright test",
+    "test:e2e:ui": "playwright test --ui"
   },
   "dependencies": {
     "@ai-sdk/openai": "^1.3.10",
@@ -86,6 +88,7 @@
   "devDependencies": {
     "firebase-tools": "^14.4.0",
     "vitest": "^3.0.5",
+    "@playwright/test": "^1.51.0",
     "@types/node": "^22",
     "@types/react": "^19",
     "@types/react-dom": "^19",
